Rename AthenzIdentityProviderImpl classes
